+++ CHANGELOG.md
@@ -10,7 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Async Zarr IO backend with non-blocking write calls
+- Different compression codec support in the ZarrBackend with `zarr_codecs` parameter
+- IO performance example
+
 ### Changed
+
+- Zarr IO Backend now uncompressed by default
 
 ### Deprecated
 
@@ -18,9 +24,25 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- DLWP output coords lead_time array to have proper shape
+
 ### Security
 
 ### Dependencies
+
+## [0.8.1] - 2025-07-07
+
+### Changed
+
+- Updated default StormCast package version to 1.0.2
+
+### Fixed
+
+- NGC filesystem from API change in version >=3.158.1 of ngcsdk
+
+### Dependencies
+
+- Removed ngcsdk dependency requirement for public NGC packages
 
 ## [0.8.0] - 2025-06-13